Implement remaining COBOL-85 gaps: 12 items via 4 parallel agents
Agent 1 — SYNCHRONIZED + COMP-1/COMP-2 IEEE 754:
- SYNCHRONIZED alignment padding in StorageLayoutComputer (2/4/8-byte boundaries)
- COMP-1: IEEE 754 single float (DecodeComp1/EncodeComp1)
- COMP-2: IEEE 754 double float (DecodeComp2/EncodeComp2)
- New IrPrimitiveType.Float32/Float64
- 3 integration tests

Agent 2 — LOCAL-STORAGE + EXTERNAL + GLOBAL:
- LOCAL-STORAGE: separate byte array in ProgramState, re-initialized on each
  Entry call via ReinitializeLocalStorage(), defaults snapshot after VALUE init
- EXTERNAL: ExternalStorage.cs with ConcurrentDictionary shared storage,
  CilEmitter redirects EXTERNAL item access to shared arrays
- GLOBAL: CBL3119 warning removed; full nested visibility deferred
  (requires nested program architecture)
- 5 integration tests

Agent 3 — File status codes + CLOSE options + READ PREVIOUS + USE:
- Status codes 02/04/14/34/39 added and wired (duplicate alternate key detection)
- CLOSE WITH LOCK: _lockedFiles enforcement, status "38" on reopen
- READ PREVIOUS: ReadDirection enum, backward iteration in IndexedFileHandler
- USE declarative execution: EmitUseDeclarative PERFORMs handler on I/O error
- 6 new lexer tokens (LOCK, NO, REEL, REMOVAL, REWIND, UNIT)
- 4 integration tests

Agent 4 — REDEFINES/RENAMES validation:
- CBL0808: REDEFINES clause ordering warning
- CBL0813: RENAMES THRU physical ordering error
- 4 unit tests

Deferred: Report Writer (XL, not NIST Nucleus), SORT external merge (production).

Tests: 260 unit + 236 integration + 60 NIST guard.

+++ src/CobolSharp.Compiler/CodeGen/RecordLayoutBuilder.cs
@@ -148,8 +148,8 @@ public sealed class RecordLayoutBuilder
                 _ => IrPrimitiveType.Int64
             },
             UsageKind.Comp3 or UsageKind.PackedDecimal => IrPrimitiveType.Decimal,
-            UsageKind.Comp1 => IrPrimitiveType.Int32, // float mapped via Int32 for now
-            UsageKind.Comp2 => IrPrimitiveType.Int64, // double mapped via Int64 for now
+            UsageKind.Comp1 => IrPrimitiveType.Float32,
+            UsageKind.Comp2 => IrPrimitiveType.Float64,
             _ => IrPrimitiveType.ByteArray
         };
     }
